backend/requirements.txt
asgiref==3.10.0
Django==5.2.7
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework-gis==1.2.0
psycopg2-binary==2.9.11
python-dotenv==1.2.1
sqlparse==0.5.3
dj-database-url
requests